+++ src/commands/project/project.ts
@@ -1,0 +1,11 @@
+import { Command } from "@cliffy/command";
+import { listCommand } from "./project-list.ts";
+import { viewCommand } from "./project-view.ts";
+
+export const projectCommand = new Command()
+  .description("Manage Linear projects")
+  .action(function () {
+    this.showHelp();
+  })
+  .command("list", listCommand)
+  .command("view", viewCommand);